Chatbot/requirements.txt
langchain
huggingface_hub
chainlit
python-dotenv
langchain-google-genai
asyncio
pandas
